plot_agent/requirements.txt
openai>=1.44
mcp>=0.1.0
pydantic>=2.7,<3
python-dotenv
pandas
matplotlib
requests